requirements.txt
django>=5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.5
django-mongodb-backend>=1.0.0
pymongo[srv]>=4.6.0
python-dotenv>=1.0.0
Pillow>=10.0.0
google-auth>=2.23.0
requests>=2.31.0
setuptools>=70.0.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
